# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Rental Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VehicleStatus = (typeof VehicleStatus)[keyof typeof VehicleStatus];

export const VehicleStatus = {
  available: "available",
  rented: "rented",
  maintenance: "maintenance",
} as const;

export interface Vehicle {
  id: number;
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  category: string;
  dailyRate: number;
  status: VehicleStatus;
  color: string;
  mileage: number;
  fuelType: string;
  transmission: string;
  seats: number;
  imageUrl?: string | null;
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateVehicleInputStatus =
  (typeof CreateVehicleInputStatus)[keyof typeof CreateVehicleInputStatus];

export const CreateVehicleInputStatus = {
  available: "available",
  rented: "rented",
  maintenance: "maintenance",
} as const;

export interface CreateVehicleInput {
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  category: string;
  dailyRate: number;
  status?: CreateVehicleInputStatus;
  color: string;
  mileage?: number;
  fuelType?: string;
  transmission?: string;
  seats?: number;
  imageUrl?: string | null;
  description?: string | null;
}

export type UpdateVehicleInputStatus =
  (typeof UpdateVehicleInputStatus)[keyof typeof UpdateVehicleInputStatus];

export const UpdateVehicleInputStatus = {
  available: "available",
  rented: "rented",
  maintenance: "maintenance",
} as const;

export interface UpdateVehicleInput {
  make?: string;
  model?: string;
  year?: number;
  licensePlate?: string;
  category?: string;
  dailyRate?: number;
  status?: UpdateVehicleInputStatus;
  color?: string;
  mileage?: number;
  fuelType?: string;
  transmission?: string;
  seats?: number;
  imageUrl?: string | null;
  description?: string | null;
}

export interface Customer {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address?: string | null;
  driverLicenseNumber: string;
  dateOfBirth?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCustomerInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address?: string | null;
  driverLicenseNumber: string;
  dateOfBirth?: string | null;
}

export interface UpdateCustomerInput {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  address?: string | null;
  driverLicenseNumber?: string;
  dateOfBirth?: string | null;
}

export type RentalStatus = (typeof RentalStatus)[keyof typeof RentalStatus];

export const RentalStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Rental {
  id: number;
  vehicleId: number;
  customerId: number;
  startDate: string;
  endDate: string;
  actualReturnDate?: string | null;
  dailyRate: number;
  totalCost?: number | null;
  status: RentalStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type RentalWithDetailsStatus =
  (typeof RentalWithDetailsStatus)[keyof typeof RentalWithDetailsStatus];

export const RentalWithDetailsStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface RentalWithDetails {
  id: number;
  vehicleId: number;
  customerId: number;
  startDate: string;
  endDate: string;
  actualReturnDate?: string | null;
  dailyRate: number;
  totalCost?: number | null;
  status: RentalWithDetailsStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  vehicle: Vehicle;
  customer: Customer;
}

export interface CreateRentalInput {
  vehicleId: number;
  customerId: number;
  startDate: string;
  endDate: string;
  dailyRate: number;
  notes?: string | null;
}

export type UpdateRentalInputStatus =
  (typeof UpdateRentalInputStatus)[keyof typeof UpdateRentalInputStatus];

export const UpdateRentalInputStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateRentalInput {
  status?: UpdateRentalInputStatus;
  actualReturnDate?: string | null;
  totalCost?: number | null;
  notes?: string | null;
}

export interface DashboardStats {
  totalVehicles: number;
  availableVehicles: number;
  rentedVehicles: number;
  maintenanceVehicles: number;
  totalCustomers: number;
  activeRentals: number;
  completedRentals: number;
  totalRevenue: number;
  monthlyRevenue: number;
}

export type ListVehiclesParams = {
  status?: ListVehiclesStatus;
  category?: string;
};

export type ListVehiclesStatus =
  (typeof ListVehiclesStatus)[keyof typeof ListVehiclesStatus];

export const ListVehiclesStatus = {
  available: "available",
  rented: "rented",
  maintenance: "maintenance",
} as const;

export type ListCustomersParams = {
  search?: string;
};

export type ListRentalsParams = {
  status?: ListRentalsStatus;
  customerId?: number;
  vehicleId?: number;
};

export type ListRentalsStatus =
  (typeof ListRentalsStatus)[keyof typeof ListRentalsStatus];

export const ListRentalsStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;