# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vehicle Rental Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: vehicles
    description: Vehicle management
  - name: customers
    description: Customer management
  - name: rentals
    description: Rental management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [available, rented, maintenance]
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Add a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleInput"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"

  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Vehicle not found
    put:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleInput"
      responses:
        "200":
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Vehicle deleted

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerInput"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Customer deleted

  /rentals:
    get:
      operationId: listRentals
      tags: [rentals]
      summary: List all rentals
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, completed, cancelled]
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
        - name: vehicleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalWithDetails"
    post:
      operationId: createRental
      tags: [rentals]
      summary: Create a new rental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRentalInput"
      responses:
        "201":
          description: Rental created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalWithDetails"

  /rentals/{id}:
    get:
      operationId: getRental
      tags: [rentals]
      summary: Get rental by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalWithDetails"
        "404":
          description: Rental not found
    put:
      operationId: updateRental
      tags: [rentals]
      summary: Update rental (e.g. return vehicle)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRentalInput"
      responses:
        "200":
          description: Rental updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalWithDetails"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [rentals]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Vehicle:
      type: object
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        category:
          type: string
        dailyRate:
          type: number
        status:
          type: string
          enum: [available, rented, maintenance]
        color:
          type: string
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        seats:
          type: integer
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - make
        - model
        - year
        - licensePlate
        - category
        - dailyRate
        - status
        - color
        - mileage
        - fuelType
        - transmission
        - seats
        - createdAt
        - updatedAt

    CreateVehicleInput:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        category:
          type: string
        dailyRate:
          type: number
        status:
          type: string
          enum: [available, rented, maintenance]
          default: available
        color:
          type: string
        mileage:
          type: integer
          default: 0
        fuelType:
          type: string
          default: Gasoline
        transmission:
          type: string
          default: Automatic
        seats:
          type: integer
          default: 5
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - make
        - model
        - year
        - licensePlate
        - category
        - dailyRate
        - color

    UpdateVehicleInput:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        category:
          type: string
        dailyRate:
          type: number
        status:
          type: string
          enum: [available, rented, maintenance]
        color:
          type: string
        mileage:
          type: integer
        fuelType:
          type: string
        transmission:
          type: string
        seats:
          type: integer
        imageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true

    Customer:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
          nullable: true
        driverLicenseNumber:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - driverLicenseNumber
        - createdAt
        - updatedAt

    CreateCustomerInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
          nullable: true
        driverLicenseNumber:
          type: string
        dateOfBirth:
          type: string
          nullable: true
      required:
        - firstName
        - lastName
        - email
        - phone
        - driverLicenseNumber

    UpdateCustomerInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
          nullable: true
        driverLicenseNumber:
          type: string
        dateOfBirth:
          type: string
          nullable: true

    Rental:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        customerId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        actualReturnDate:
          type: string
          nullable: true
        dailyRate:
          type: number
        totalCost:
          type: number
          nullable: true
        status:
          type: string
          enum: [active, completed, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vehicleId
        - customerId
        - startDate
        - endDate
        - dailyRate
        - status
        - createdAt
        - updatedAt

    RentalWithDetails:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        customerId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        actualReturnDate:
          type: string
          nullable: true
        dailyRate:
          type: number
        totalCost:
          type: number
          nullable: true
        status:
          type: string
          enum: [active, completed, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        customer:
          $ref: "#/components/schemas/Customer"
      required:
        - id
        - vehicleId
        - customerId
        - startDate
        - endDate
        - dailyRate
        - status
        - createdAt
        - updatedAt
        - vehicle
        - customer

    CreateRentalInput:
      type: object
      properties:
        vehicleId:
          type: integer
        customerId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        dailyRate:
          type: number
        notes:
          type: string
          nullable: true
      required:
        - vehicleId
        - customerId
        - startDate
        - endDate
        - dailyRate

    UpdateRentalInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, cancelled]
        actualReturnDate:
          type: string
          nullable: true
        totalCost:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true

    DashboardStats:
      type: object
      properties:
        totalVehicles:
          type: integer
        availableVehicles:
          type: integer
        rentedVehicles:
          type: integer
        maintenanceVehicles:
          type: integer
        totalCustomers:
          type: integer
        activeRentals:
          type: integer
        completedRentals:
          type: integer
        totalRevenue:
          type: number
        monthlyRevenue:
          type: number
      required:
        - totalVehicles
        - availableVehicles
        - rentedVehicles
        - maintenanceVehicles
        - totalCustomers
        - activeRentals
        - completedRentals
        - totalRevenue
        - monthlyRevenue