# lib/api-zod/src/generated/types/createRentalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Rental Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRentalInput {
  vehicleId: number;
  customerId: number;
  startDate: string;
  endDate: string;
  dailyRate: number;
  notes?: string | null;
}