# lib/api-zod/src/generated/types/createVehicleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Rental Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateVehicleInputStatus } from "./createVehicleInputStatus";

export interface CreateVehicleInput {
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  category: string;
  dailyRate: number;
  status?: CreateVehicleInputStatus;
  color: string;
  mileage?: number;
  fuelType?: string;
  transmission?: string;
  seats?: number;
  imageUrl?: string | null;
  description?: string | null;
}