# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Rental Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address?: string | null;
  driverLicenseNumber: string;
  dateOfBirth?: string | null;
  createdAt: string;
  updatedAt: string;
}